WebProject/requirements.txt
flask
flask-sqlalchemy
flask-smorest
python-dotenv
Marshmallow
flask-pymongo
flask-bcrypt 
pyjwt